pyproject.toml
[tool.poetry]
name = "fitness-ai-chatbot"
version = "0.1.0"
description = ""
authors = ["RanceCL <roderickrwhite2022@gmail.com>"]
readme = "README.md"
packages = [{include = "fitness_ai_chatbot"}]

[tool.poetry.dependencies]
python = "3.12.8"
langgraph = "^0.2.67"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"
flask-cors = "^5.0.0"
psycopg2 = "^2.9.10"
email-validator = "^2.2.0"
ipykernel = "^6.29.5"
langgraph-checkpoint = "^2.0.10"
langgraph-checkpoint-postgres = "^2.0.13"
psycopg-binary = "^3.2.4"
pydantic = "^2.10.6"
langchain-openai = "^0.3.2"
ortools = "^9.11.4210"
pulp = "^2.9.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
